Chore(pyproject.toml): Update keywords and versions to reflect what is supported. Testing/Refactor: Change names of models used for testing to avoid pytest confusion Testing(warnings): Ignore from hypothesis Feat(Justfile): Add mypy typecheck for test Refactor(model): Merged into a single python file Chore(version): MAJOR
