+++ pydantic_numpy/__init__.py
@@ -1,2 +1,4 @@
 from pydantic_numpy.helper.annotation import np_array_pydantic_annotated_typing
 from pydantic_numpy.typing.n_dimensional import *
+
+__all__ = ["np_array_pydantic_annotated_typing", "model", "typing"]
