--- pydantic_numpy/model/__init__.py
@@ -1,2 +1,0 @@
-from pydantic_numpy.model.multi_array import MultiArrayNumpyFile
-from pydantic_numpy.model.np_model import NumpyModel
