+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pydantic_numpy"
-version = "3.1.5"
+version = "4.0.0"
 description = "Pydantic Model integration of the NumPy array"
 authors = ["Can H. Tartanoglu", "Christoph Heindl"]
 maintainers = ["Can H. Tartanoglu <python@rotas.mozmail.com>"]
@@ -8,7 +8,7 @@ readme = "README.md"
 homepage = "https://github.com/caniko/pydantic-numpy"
 license = "BSD-4"
 
-keywords = ["pydantic", "numpy", "typing"]
+keywords = ["pydantic", "numpy", "typing", "validation"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Topic :: Software Development :: Libraries :: Python Modules",
@@ -40,6 +40,12 @@ ruff = "^0.0.285"
 [tool.poetry.group.typecheck.dependencies]
 mypy = "*"
 
+[tool.pytest.ini_options]
+filterwarnings = [
+    "ignore::hypothesis.errors.NonInteractiveExampleWarning",
+    "ignore:invalid value encountered in multiply:RuntimeWarning",
+]
+
 [tool.black]
 line-length = 120
 target-version = ["py311"]
@@ -50,7 +56,7 @@ profile = "black"
 [tool.ruff]
 line-length = 120
 ignore-init-module-imports = true
-ignore = ["F401", "F403", "F405"]
+ignore = ["F403", "F405"]
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
