+++ tests/test_np_model.py
@@ -5,23 +5,17 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from pydantic_numpy.model import NumpyModel
-from pydantic_numpy.model.np_model import model_agnostic_load
+from pydantic_numpy.model import model_agnostic_load
 from pydantic_numpy.typing import NpNDArray
 from tests.helper.cache import cached_hyp_array
+from tests.model import (
+    NpNDArrayModelWithNonArray,
+    NpNDArrayModelWithNonArrayWithArbitrary,
+)
 
 TEST_MODEL_OBJECT_ID = "test"
 OTHER_TEST_MODEL_OBJECT_ID = "other_test"
 NON_ARRAY_VALUE = 5
-
-
-class NumpyModelForTest(NumpyModel):
-    array: NpNDArray
-    non_array: int
-
-
-class TestWithArbitraryForTest(NumpyModelForTest, arbitrary_types_allowed=True):
-    my_arbitrary_slice: slice
 
 
 def _create_example_array():
@@ -29,7 +23,7 @@ def _create_example_array():
 
 
 def _numpy_model():
-    return NumpyModelForTest(array=_create_example_array(), non_array=NON_ARRAY_VALUE)
+    return NpNDArrayModelWithNonArray(array=_create_example_array(), non_array=NON_ARRAY_VALUE)
 
 
 @pytest.fixture
@@ -40,7 +34,7 @@ def numpy_model():
 @pytest.fixture(
     params=[
         _numpy_model(),
-        TestWithArbitraryForTest(
+        NpNDArrayModelWithNonArrayWithArbitrary(
             array=_create_example_array(), non_array=NON_ARRAY_VALUE, my_arbitrary_slice=slice(0, 10)
         ),
     ]
@@ -51,33 +45,33 @@ def numpy_model_with_arbitrary(request):
 
 if os.name != "nt":
 
-    def test_io_yaml(numpy_model: NumpyModel) -> None:
+    def test_io_yaml(numpy_model: NpNDArrayModelWithNonArray) -> None:
         with tempfile.TemporaryDirectory() as tmp_dirname:
             numpy_model.dump(tmp_dirname, TEST_MODEL_OBJECT_ID)
             assert numpy_model.load(tmp_dirname, TEST_MODEL_OBJECT_ID) == numpy_model
 
-    def test_io_compressed_pickle(numpy_model_with_arbitrary: NumpyModel) -> None:
+    def test_io_compressed_pickle(numpy_model_with_arbitrary: NpNDArrayModelWithNonArray) -> None:
         with tempfile.TemporaryDirectory() as tmp_dirname:
             numpy_model_with_arbitrary.dump(tmp_dirname, TEST_MODEL_OBJECT_ID, pickle=True)
             assert numpy_model_with_arbitrary.load(tmp_dirname, TEST_MODEL_OBJECT_ID) == numpy_model_with_arbitrary
 
-    def test_io_pickle(numpy_model_with_arbitrary: NumpyModel) -> None:
+    def test_io_pickle(numpy_model_with_arbitrary: NpNDArrayModelWithNonArray) -> None:
         with tempfile.TemporaryDirectory() as tmp_dirname:
             numpy_model_with_arbitrary.dump(tmp_dirname, TEST_MODEL_OBJECT_ID, pickle=True, compress=False)
             assert numpy_model_with_arbitrary.load(tmp_dirname, TEST_MODEL_OBJECT_ID) == numpy_model_with_arbitrary
 
-    def test_typing_json_dump(numpy_model: NumpyModel):
+    def test_typing_json_dump(numpy_model: NpNDArrayModelWithNonArray):
         assert numpy_model.model_dump_json() == '{"array":"%s","non_array":%s}' % (
             np.array2string(numpy_model.array),
             NON_ARRAY_VALUE,
         ), ""
 
     def test_model_agnostic_load():
-        class NumpyModelAForTest(NumpyModel):
+        class NumpyModelAForTest(NpNDArrayModelWithNonArray):
             array: NpNDArray
             non_array: int
 
-        class NumpyModelBForTest(NumpyModel):
+        class NumpyModelBForTest(NpNDArrayModelWithNonArray):
             array: NpNDArray
             non_array: int
