+++ tests/test_pydantic.py
@@ -1,10 +1,5 @@
-from pydantic_numpy.model import NumpyModel
-from pydantic_numpy.typing import NpNDArray
-
-
-class TestModel(NumpyModel):
-    array: NpNDArray
+from tests.model import NpNDArrayModel
 
 
 def test_model_json_schema():
-    assert TestModel.model_json_schema()
+    assert NpNDArrayModel.model_json_schema()
